Minor revision

+++ models/utils.py
@@ -10,12 +10,12 @@ from typing import Optional, List
 def cat2ohe(x, a, num_node_types, num_edge_types):
     x = torch.nn.functional.one_hot(x.to(torch.int64), num_classes=num_node_types)
     a = torch.nn.functional.one_hot(a.to(torch.int64), num_classes=num_edge_types)
-    return x, a
+    return x.to(torch.float), a.to(torch.float)
 
 def ohe2cat(x, a):
     x = torch.argmax(x, dim=-1)
     a = torch.argmax(a, dim=-1)
-    return x, a
+    return x.to(torch.float), a.to(torch.float)
 
 def set_diagonal(a, device, val=-torch.inf):
     mask = torch.eye(a.shape[1], dtype=bool, device=device)
